[Physics] Moved `SAT` and moved functionality to `SATNarrowPhase` #change
